Sample code showing the use of TestMediators
Mocks and Mocking are a solution to a certain class of problems. The
problems Mocking solves are valid and important. However, I don't
believe the way we achieve these solutions (via Mocking) is the correct
solution. In other words the means to the end is what I have a problem
with. TestMediators solve the same problem as Mocks, but the way we get
there (to the solution) is very different and I prefer using
TestMediators to Mocking as a result. The project here provides samples
code for a simple system and a couple of unit tests that use a
TestMediator.